Output File Function
Coded in the output file function

This also updated the algorithm  and Node Class. I added a "function cost" member to hold f(n), added a "numberMoved"  member to hold the number  that moved with this node, made it so the algorithm keeps track of the numberMoved for each node and the algorithm updates a new global variable called "searchedNodes", which for now I made equal to the closed list.